dashboard-cloud/requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
websockets